requirements.txt
asgiref==3.5.2
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cloudinary==1.29.0
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.14
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-filter==22.1
django-model-utils==4.2.0
django-summernote==0.8.20.0
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.10
webencodings==0.5.1
